kanban-backend/kanban_backend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
awscli==1.29.85
boto3==1.28.85
botocore==1.31.85
colorama==0.4.4
Django==4.2.7
django-cors-headers==4.3.1
docutils==0.16
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
jmespath==1.0.1
promise==2.3
pyasn1==0.5.0
python-dateutil==2.8.2
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
urllib3==2.0.7
